Create loci_filtering.py
This Python script is designed to filter and move FASTA sequence alignments based on specific criteria, and it is particularly useful for genetic and bioinformatic analyses. It leverages the BioPython library and the AMAS package to process a set of sequence alignment files in a specified directory, filtering them according to criteria related to GC content, missing sites, and parsimony informative sites. The selected alignments meeting these criteria are then moved to a new directory.